Share more code (again)
