+++ src/Essentials/src/Barometer/Barometer.ios.watchos.cs
@@ -4,78 +4,34 @@ using Foundation;
 
 namespace Microsoft.Maui.Essentials.Implementations
 {
-	public class BarometerImplementation : IBarometer
+	public partial class BarometerImplementation : IBarometer
 	{
-		public bool IsSupported
-			=> CMAltimeter.IsRelativeAltitudeAvailable;
-
-		public event EventHandler<BarometerChangedEventArgs> ReadingChanged;
-
 		static CMAltimeter altitudeManager;
 
-		bool UseSyncContext => SensorSpeed == SensorSpeed.Default || SensorSpeed == SensorSpeed.UI;
+		bool PlatformIsSupported
+			=> CMAltimeter.IsRelativeAltitudeAvailable;
 
-		public SensorSpeed SensorSpeed { get; private set; } = SensorSpeed.Default;
-
-		public bool IsMonitoring { get; private set; }
-
-		public void Start(SensorSpeed sensorSpeed)
+		void PlatformStart(SensorSpeed sensorSpeed)
 		{
-			if (!IsSupported)
-				throw new FeatureNotSupportedException();
+			altitudeManager = new CMAltimeter();
+			altitudeManager.StartRelativeAltitudeUpdates(Platform.GetCurrentQueue(), LocationManagerUpdatedHeading);
 
-			if (IsMonitoring)
-				throw new InvalidOperationException("Barometer has already been started.");
-
-			IsMonitoring = true;
-			SensorSpeed = sensorSpeed;
-
-			try
+			void LocationManagerUpdatedHeading(CMAltitudeData e, NSError error)
 			{
-				altitudeManager = new CMAltimeter();
-				altitudeManager.StartRelativeAltitudeUpdates(Platform.GetCurrentQueue(), LocationManagerUpdatedHeading);
+				var reading = new BarometerData(UnitConverters.KilopascalsToHectopascals(e.Pressure.DoubleValue));
 
-				void LocationManagerUpdatedHeading(CMAltitudeData e, NSError error)
-				{
-					var reading = new BarometerData(UnitConverters.KilopascalsToHectopascals(e.Pressure.DoubleValue));
-
-					if (UseSyncContext)
-						MainThread.BeginInvokeOnMainThread(() => ReadingChanged?.Invoke(this, new BarometerChangedEventArgs(reading)));
-					else
-						ReadingChanged?.Invoke(this, new BarometerChangedEventArgs(reading));
-				}
-			}
-			catch
-			{
-				IsMonitoring = false;
-				throw;
+				RaiseReadingChanged(reading);
 			}
 		}
 
-		public void Stop()
+		void PlatformStop()
 		{
-			if (!IsSupported)
-				throw new FeatureNotSupportedException();
-
-			if (!IsMonitoring)
-				return;
-
-			IsMonitoring = false;
-
 			if (altitudeManager == null)
 				return;
 
-			try
-			{
-				altitudeManager.StopRelativeAltitudeUpdates();
-				altitudeManager.Dispose();
-				altitudeManager = null;
-			}
-			catch
-			{
-				IsMonitoring = true;
-				throw;
-			}
+			altitudeManager.StopRelativeAltitudeUpdates();
+			altitudeManager.Dispose();
+			altitudeManager = null;
 		}
 	}
 }
