+++ src/Essentials/src/Barometer/Barometer.tizen.cs
@@ -6,74 +6,26 @@ namespace Microsoft.Maui.Essentials.Implementations
 {
 	public class BarometerImplementation : IBarometer
 	{
-		bool UseSyncContext => SensorSpeed == SensorSpeed.Default || SensorSpeed == SensorSpeed.UI;
-
-		public SensorSpeed SensorSpeed { get; private set; } = SensorSpeed.Default;
-
-		public event EventHandler<BarometerChangedEventArgs> ReadingChanged;
-
 		static TizenBarometerSensor DefaultSensor
 			=> (TizenBarometerSensor)Platform.GetDefaultSensor(SensorType.Barometer);
 
-		public bool IsSupported
+		bool PlatformIsSupported
 			=> TizenBarometerSensor.IsSupported;
 
-		public bool IsMonitoring { get; private set; }
-
-		public void Start(SensorSpeed sensorSpeed)
+		void PlatformStart(SensorSpeed sensorSpeed)
 		{
-			if (!IsSupported)
-				throw new FeatureNotSupportedException();
-
-			if (IsMonitoring)
-				throw new InvalidOperationException("Barometer has already been started.");
-
-			IsMonitoring = true;
-			SensorSpeed = sensorSpeed;
-
-			try
-			{
-				DefaultSensor.Interval = sensorSpeed.ToPlatform();
-				DefaultSensor.DataUpdated += DataUpdated;
-				DefaultSensor.Start();
-			}
-			catch
-			{
-				IsMonitoring = false;
-				throw;
-			}
+			DefaultSensor.Interval = sensorSpeed.ToPlatform();
+			DefaultSensor.DataUpdated += DataUpdated;
+			DefaultSensor.Start();
 		}
 
 		void DataUpdated(object sender, PressureSensorDataUpdatedEventArgs e)
+			=> RaiseDataChanged(new BarometerData(e.Pressure));
+
+		void PlatformStop()
 		{
-			var data = new BarometerData(e.Pressure);
-
-			if (UseSyncContext)
-				MainThread.BeginInvokeOnMainThread(() => ReadingChanged?.Invoke(this, new BarometerChangedEventArgs(data)));
-			else
-				ReadingChanged?.Invoke(this, new BarometerChangedEventArgs(data));
-		}
-
-		public void Stop()
-		{
-			if (!IsSupported)
-				throw new FeatureNotSupportedException();
-
-			if (!IsMonitoring)
-				return;
-
-			IsMonitoring = false;
-
-			try
-			{
-				DefaultSensor.DataUpdated -= DataUpdated;
-				DefaultSensor.Stop();
-			}
-			catch
-			{
-				IsMonitoring = true;
-				throw;
-			}
+			DefaultSensor.DataUpdated -= DataUpdated;
+			DefaultSensor.Stop();
 		}
 	}
 }
